COM-1755 export traineesTable
