+++ src/core/components/courses/ProfileFollowUp.vue
@@ -1,116 +1,46 @@
 <template>
   <div>
-    <div v-if="course.format === STRICTLY_E_LEARNING" class="q-mb-xl">
-      <p class="text-weight-bold">Rapport d'utilisation</p>
-      <div class="row">
-        <div class="col-md-6 col-xs-12">
-          <q-card flat class="q-pa-md right-stats">
-            <div class="text-weight-bold q-mb-sm">Vue globale</div>
-            <div class="row">
-              <div class="col-3 text-right q-pr-md column justify-around">
-                <ni-e-learning-indicator :indicator="traineesOnGoingCount" />
-                <ni-e-learning-indicator :indicator="traineesFinishedCount" />
-              </div>
-              <div class="col-9 column justify-around">
-                <div>apprenant{{ traineesOnGoingCount > 1 ? 's' : '' }} en cours</div>
-                <div>apprenant{{ traineesFinishedCount > 1 ? 's' : '' }} ayant terminé</div>
-              </div>
+    <p class="text-weight-bold">Rapport d'utilisation</p>
+    <div class="row">
+      <div class="col-md-6 col-xs-12">
+        <q-card flat class="q-pa-md right-stats">
+          <div class="text-weight-bold q-mb-sm">Vue globale</div>
+          <div class="row">
+            <div class="col-3 text-right q-pr-md column justify-around">
+              <ni-e-learning-indicator :indicator="traineesOnGoingCount" />
+              <ni-e-learning-indicator :indicator="traineesFinishedCount" />
             </div>
-          </q-card>
-        </div>
+            <div class="col-9 column justify-around">
+              <div>apprenant{{ traineesOnGoingCount > 1 ? 's' : '' }} en cours</div>
+              <div>apprenant{{ traineesFinishedCount > 1 ? 's' : '' }} ayant terminé</div>
+            </div>
+          </div>
+        </q-card>
       </div>
     </div>
-    <p class="text-weight-bold">Progression des participants</p>
-    <q-card>
-      <ni-expanding-table :data="learners" :columns="columns" :pagination="pagination" :hide-bottom="false">
-        <template #row="{ props }">
-          <q-td v-for="col in props.cols" :key="col.name" :props="props">
-            <template v-if="col.name === 'progress'">
-              <ni-progress class="progress" :value="col.value" />
-            </template>
-            <template v-else-if="col.name === 'expand'">
-              <q-icon :name="props.expand ? 'expand_less' : 'expand_more'" />
-            </template>
-            <template v-else>
-              <div class="name" @click.stop="goToLearnerProfile(props.row)">{{ col.value }}</div>
-            </template>
-          </q-td>
-        </template>
-        <template #expanding-row="{ props }">
-          <q-td colspan="100%">
-            <div v-for="(step, stepIndex) in props.row.steps" :key="step._id" :props="props"
-              class="q-ma-sm expanding-table-expanded-row">
-              <div>
-                <q-icon :name="step.type === E_LEARNING ? 'stay_current_portrait' : 'mdi-teach'" />
-                {{ stepIndex + 1 }} - {{ step.name }}
-              </div>
-              <div class="expanding-table-progress-container">
-                <ni-progress class="expanding-table-sub-progress" :value="step.progress" />
-              </div>
-            </div>
-          </q-td>
-        </template>
-      </ni-expanding-table>
-    </q-card>
+    <trainee-follow-up-table :learners.sync="learners" :profile-id="profileId" class="q-mt-xl" />
   </div>
 </template>
 
 <script>
-import { mapState, mapGetters } from 'vuex';
-import Courses from '@api/Courses';
-import ExpandingTable from '@components/table/ExpandingTable';
-import Progress from '@components/CourseProgress';
-import { NotifyNegative } from '@components/popup/notify';
-import { sortStrings, formatIdentity } from '@helpers/utils';
-import { STRICTLY_E_LEARNING, E_LEARNING } from '@data/constants.js';
 import ELearningIndicator from '@components/courses/ELearningIndicator';
+import TraineeFollowUpTable from '@components/courses/TraineeFollowUpTable';
 
 export default {
   name: 'ProfileFollowUp',
   components: {
-    'ni-expanding-table': ExpandingTable,
-    'ni-progress': Progress,
     'ni-e-learning-indicator': ELearningIndicator,
+    'trainee-follow-up-table': TraineeFollowUpTable,
   },
   props: {
     profileId: { type: String, required: true },
   },
   data () {
-    const isClientInterface = !/\/ad\//.test(this.$router.currentRoute.path);
-
     return {
-      tableLoading: false,
-      columns: [
-        {
-          name: 'name',
-          label: 'Nom',
-          field: 'identity',
-          format: value => (value ? value.fullName : ''),
-          align: 'left',
-          sortable: true,
-          sort: (a, b) => sortStrings(a.lastname, b.lastname),
-          style: 'width: 70%',
-        },
-        {
-          name: 'progress',
-          label: 'Progression',
-          field: 'progress',
-          align: 'center',
-          sortable: true,
-          style: 'min-width: 150px; width: 20%',
-        },
-        { name: 'expand', label: '', field: '' },
-      ],
       learners: [],
-      pagination: { sortBy: 'name', ascending: true, page: 1, rowsPerPage: 15 },
-      E_LEARNING,
-      STRICTLY_E_LEARNING,
-      isClientInterface,
     };
   },
   computed: {
-    ...mapState('course', ['course']),
-    ...mapGetters({ company: 'main/getCompany' }),
     traineesOnGoingCount () {
       return this.learners.filter(l => l.progress !== 1).length;
     },
@@ -118,50 +48,5 @@ export default {
       return this.learners.length - this.traineesOnGoingCount;
     },
   },
-  async created () {
-    await this.getLearnersList();
-  },
-  methods: {
-    goToLearnerProfile (row) {
-      const name = this.isClientInterface ? 'ni courses learners info' : 'ni users learners info';
-      this.$router.push({ name, params: { learnerId: row._id, defaultTab: 'courses' } });
-    },
-    formatRow (trainee) {
-      const formattedName = formatIdentity(trainee.identity, 'FL');
-
-      return {
-        _id: trainee._id,
-        identity: { ...trainee.identity, fullName: formattedName },
-        progress: trainee.progress,
-        steps: trainee.steps,
-      };
-    },
-    async getLearnersList () {
-      try {
-        this.tableLoading = true;
-        const course = await Courses.getFollowUp(
-          this.profileId,
-          this.isClientInterface ? { company: this.company._id } : null
-        );
-
-        if (course) this.learners = Object.freeze(course.trainees.map(this.formatRow));
-      } catch (e) {
-        console.error(e);
-        NotifyNegative('Erreur lors de la récupération des apprenants');
-        this.learners = [];
-      } finally {
-        this.tableLoading = false;
-      }
-    },
-  },
 };
 </script>
-
-<style lang="stylus" scoped>
-  .progress
-    width: 100%
-  .name
-    width: fit-content;
-    text-decoration: underline
-    color: $primary
-</style>
